v0.17.1.0 add Rank Drugs analysis + bugfixes
New report: Rank Drugs for a Sample (step 8.75)
Requires the provided target-drug_relationships_2020-02-27.tsv reference file (md5 07ac00f30632915bfb2089cd500de545).
This script generates a plot and table showing expression of all drug targets identified as leads in the CKCC2 project,
sorted by highest expression in the focus sample. Results are added to the end of Summary.html

Minor Enhancements:
For each pandisease cohort, when sample counts by disease are reported, we now also report what percentage of samples of
that disease are in the cohort. Applies to first-degree and first-and-second-degree cohorts only, as other cohorts are disease-based
and will thus always be 100%. For example, if there are 30 AML samples in the compendium and 10 of them are in the cohort, this will
report (33%).

The annotations.json file now includes a field optional_rin_score for displaying the RIN of a focus sample in the summary and slides documents.

A virtual "sticky note" has been added to the summmary to improve printing experience.

Headers & footers improved in summary.

GSEA pathway names are now broken with spaces rather than underscores for better wrapping.

Improved wording on message when a pandisease cohort is too small.

Better display of blank disease names.

Bugfixes:
QC report notices in log.txt no longer claim to be errors.
RIN range now maxes out at 10.0 rather than 10.1.
Fixed bug highlighting the wrong pandisease cohort under some conditions.